+++ lib/pages/chat/events/image_bubble.dart
@@ -33,6 +33,7 @@ class ImageBubble extends StatelessWidget {
   final Timeline? timeline;
   final InlineSpan? trailingSpan;
   final MessageLayout layout;
+  final bool selectable;
 
   final bool ownMessage;
   final bool previousEventSameSender;
@@ -67,6 +68,7 @@ class ImageBubble extends StatelessWidget {
     this.onRevealHiddenMedia,
     this.contentWarning,
     this.trailingSpan,
+    this.selectable = true,
     required this.layout,
     super.key,
   });
@@ -337,7 +339,7 @@ class ImageBubble extends StatelessWidget {
                       AppSettings.messageFontSize.value,
                   decoration: .none,
                 ),
-                selectable: true,
+                selectable: selectable,
                 onOpen: (url) => UrlLauncher(context, url.url).launchUrl(),
                 onCopy: () {
                   Clipboard.setData(ClipboardData(text: event.body));