+++ lib/pages/chat/events/message_bubble.dart
@@ -52,6 +52,7 @@ class MessageBubble extends StatefulWidget {
   final bool singleSelected;
   final Thread? thread;
   final bool hasBeenRead;
+  final bool selectable;
   final bool? exampleMessage;
 
   const MessageBubble(
@@ -75,6 +76,7 @@ class MessageBubble extends StatefulWidget {
     this.animateIn = false,
     this.wallpaperMode = false,
     required this.onMention,
+    this.selectable = true,
     this.scrollController,
     this.chatController,
     required this.colors,
@@ -587,7 +589,12 @@ class _MessageBubbleState extends State<MessageBubble> {
                 onTapDown: (details) => _tapPosition = details.globalPosition,
                 onSecondaryTapDown: (details) =>
                     _tapPosition = details.globalPosition,
-                onTap: () => _handleQuickActionTap(),
+                onTap: widget.longPressSelect
+                    ? () {
+                        HapticFeedback.heavyImpact();
+                        widget.onSelect(event, _tapPosition);
+                      }
+                    : () => _handleQuickActionTap(),
                 onLongPress: () {
                   if (PlatformInfos.isMobile) {
                     widget.onSelect(event, _tapPosition);
@@ -681,7 +688,15 @@ class _MessageBubbleState extends State<MessageBubble> {
                                   HapticFeedback.heavyImpact();
                                   widget.onSelect(event, _tapPosition);
                                 }
-                              : null,
+                              : () {
+                                  if (PlatformInfos.isMobile) {
+                                    _handleQuickActionTap();
+                                  }
+                                },
+                          onSecondaryTapDown: (details) =>
+                              _tapPosition = details.globalPosition,
+                          onSecondaryTap: () =>
+                              widget.onSelect(event, _tapPosition),
                           child: Material(
                             color: noBubble
                                 ? Colors.transparent
@@ -784,17 +799,7 @@ class _MessageBubbleState extends State<MessageBubble> {
                                                     previousEventSameSender,
                                                 ownMessage: ownMessage,
                                                 layout: .bubbles,
-                                                selectable:
-                                                    PlatformInfos.isMobile
-                                                    ? widget.longPressSelect
-                                                    : true,
-                                                onSecondaryTap: (position) {
-                                                  _tapPosition = position;
-                                                  widget.onSelect(
-                                                    event,
-                                                    position,
-                                                  );
-                                                },
+                                                selectable: widget.selectable,
                                                 trailingSpan: useInlineStatus
                                                     ? inlineStatusPlaceholder
                                                     : null,