+++ lib/utils/uia_request_manager.dart
@@ -42,7 +42,7 @@ extension UiaRequestManager on MatrixState {
           if (input == null || input.isEmpty) {
             return uiaRequest.cancel();
           }
-          return uiaRequest.completeStage(
+          return await uiaRequest.completeStage(
             AuthenticationPassword(
               session: uiaRequest.session,
               password: input,
@@ -72,11 +72,11 @@ extension UiaRequestManager on MatrixState {
                 okLabel: l10n.iHaveClickedOnLink,
                 cancelLabel: l10n.cancel,
               )) {
-            return uiaRequest.completeStage(auth);
+            return await uiaRequest.completeStage(auth);
           }
           return uiaRequest.cancel();
         case AuthenticationTypes.dummy:
-          return uiaRequest.completeStage(
+          return await uiaRequest.completeStage(
             AuthenticationData(
               type: AuthenticationTypes.dummy,
               session: uiaRequest.session,
@@ -111,7 +111,7 @@ extension UiaRequestManager on MatrixState {
           await completer.future;
           listener.dispose();
 
-          return uiaRequest.completeStage(
+          return await uiaRequest.completeStage(
             AuthenticationData(session: uiaRequest.session),
           );
       }