+++ windows/flutter/generated_plugins.cmake
@@ -16,8 +16,6 @@ list(APPEND FLUTTER_PLUGIN_LIST
   hotkey_manager_windows
   livekit_client
   local_auth_windows
-  media_kit_libs_windows_video
-  media_kit_video
   pasteboard
   permission_handler_windows
   record_windows